Fix #14

+++ Minty/M_C_Camera.h
@@ -14,10 +14,10 @@ namespace minty
 		/// <summary>
 		/// The offset of the Camera when rendering.
 		/// </summary>
-		int offsetX, offsetY;
+		float offsetX, offsetY;
 
 		void setCenter(float const x, float const y, Screen* const screen);
 
-		void setOffset(int const x, int const y);
+		void setOffset(float const x, float const y);
 	};
 }
